feat(plan): consolidate Board/Grid/Outline into one Structure sub-tab
Plan opened onto five flat sub-tabs (Overview, Beats, Board, Grid,
Outline) with no implied order — overwhelming for newcomers. Board,
Grid, and the outline tree are all views of the same scene structure,
so fold them under one Structure sub-tab with a Tree | Board | Grid
switcher. Plan is now Overview -> Beats -> Structure.

- New thin wrapper StructurePanel (src/plan/structure-panel.ts) owns the
  switcher + per-view hint and delegates to the existing panels unchanged
- Rename the outline tree StructurePanel -> OutlinePanel
  (src/plan/outline-panel.ts) to free the name
- Host view exposes openPlanStructure(view); open-board / open-plot-grid
  commands keep their ids and deep-link into the Board/Grid views
- Hint keys stay frozen (plan/board|grid|outline) so dismissals persist
- New shared renderEmptyStateAction primitive powers smarter empty states
  that point to the natural next step (e.g. empty Structure -> Beats)

Verified in the dev vault: 3 sub-tabs, view switching + memory, deep-link
commands, hint dismissal persistence. typecheck + 469 tests + lint clean.
