+++ src/help/help-content.ts
@@ -152,16 +152,18 @@ export const HINTS: Record<string, HintEntry> = {
     body: (el) => {
       p(
         el,
-        "Organise your book as a tree: acts hold chapters, chapters hold scenes. Drag a " +
-          "scene into a chapter (or a chapter into an act), and drop above or below a row " +
-          "to reorder. This tree is the source of truth — arranging it also sets the " +
-          "manuscript's reading order, so each chapter's scenes stay together."
+        "The Tree view of Structure: organise your book as acts holding chapters holding " +
+          "scenes. Drag a scene into a chapter (or a chapter into an act), and drop above " +
+          "or below a row to reorder. This tree is the source of truth — arranging it also " +
+          "sets the manuscript's reading order, so each chapter's scenes stay together. " +
+          "The Board and Grid views (top of the tab) show the same scenes a different way."
       );
       p(
         el,
-        "It comes last in Plan on purpose. The usual flow is to sketch your Beats first, " +
-          "spin scenes off those beats, and draft them in Write — structure can wait. " +
-          "Once you have real scenes, come back here to group them into acts and chapters."
+        "Structure comes last in Plan on purpose. The usual flow is to sketch your Beats " +
+          "first, spin scenes off those beats, and draft them in Write — structure can " +
+          "wait. Once you have real scenes, come back here to group them into acts and " +
+          "chapters."
       );
       steps(el, [
         "Nesting is optional — a scene can sit outside any chapter and a chapter outside any act (the “Chapters with no act” / “Unassigned scenes” buckets).",
@@ -316,21 +318,23 @@ export const HELP_SECTIONS: HelpSection[] = [
   {
     phase: "Plan",
     icon: "compass",
-    summary: "Overview fields, the beat sheet, the status board, the plot grid, and the story outline.",
+    summary: "Overview fields, the beat sheet, and Structure (outline tree, board, and plot grid).",
     body: (el) => {
       p(
         el,
         "Overview holds novel-level fields (logline, theme, genre) and long-form prose " +
-          "(synopsis, three-act sketch). Beats, Board, Grid, and Outline are detailed below."
+          "(synopsis, three-act sketch). Beats is the beat sheet. Structure combines the " +
+          "outline tree, the Kanban board, and the plot grid behind one Tree | Board | " +
+          "Grid switcher — three views of the same scenes. All are detailed below."
       );
       el.createEl("h4", { text: HINTS["plan/beats"].title });
       HINTS["plan/beats"].body(el);
+      el.createEl("h4", { text: HINTS["plan/outline"].title });
+      HINTS["plan/outline"].body(el);
       el.createEl("h4", { text: HINTS["plan/board"].title });
       HINTS["plan/board"].body(el);
       el.createEl("h4", { text: HINTS["plan/grid"].title });
       HINTS["plan/grid"].body(el);
-      el.createEl("h4", { text: HINTS["plan/outline"].title });
-      HINTS["plan/outline"].body(el);
     },
   },
   {
