+++ src/views/inkswell-view.ts
@@ -19,10 +19,8 @@ import { NewDraftModal } from "./drafts-modal";
 import { CodexPanel } from "../codex/codex-panel";
 import { AnalysisPanel } from "../insight/analysis-panel";
 import { BeatPanel } from "../outliner/beat-panel";
-import { BoardPanel } from "../outliner/board-panel";
 import { OverviewPanel } from "../plan/overview-panel";
-import { PlotGridPanel } from "../plan/plotgrid-panel";
-import { StructurePanel } from "../plan/structure-panel";
+import { StructurePanel, StructureView } from "../plan/structure-panel";
 import { ProjectStats } from "../projects/project-stats";
 import { ProjectStore } from "../projects/project-store";
 import { ChecklistPanel } from "./publish/checklist-panel";
@@ -56,8 +54,6 @@ export class InkswellView extends ItemView {
   private explorer: ExplorerPanel;
   private overview: OverviewPanel;
   private beats: BeatPanel;
-  private board: BoardPanel;
-  private plotGrid: PlotGridPanel;
   private structure: StructurePanel;
   private codex: CodexPanel;
   private write: WritePanel;
@@ -114,8 +110,6 @@ export class InkswellView extends ItemView {
     });
     this.overview = new OverviewPanel(this.app, plugin, store, plugin.activeProject);
     this.beats = new BeatPanel(this.app, plugin, store, plugin.activeProject);
-    this.board = new BoardPanel(this.app, plugin, store, plugin.activeProject);
-    this.plotGrid = new PlotGridPanel(this.app, plugin, store, plugin.activeProject);
     this.structure = new StructurePanel(this.app, plugin, store, plugin.activeProject);
     this.codex = new CodexPanel(this.app, plugin);
     // On phones a codex row tap drills into a single-column detail screen; on
@@ -420,6 +414,12 @@ export class InkswellView extends ItemView {
     this.setMode("write");
   }
 
+  /** Open Plan → Structure on a specific view (deep link for board/grid commands). */
+  openPlanStructure(view: StructureView): void {
+    this.structure.setView(view);
+    this.setMode("plan", "structure");
+  }
+
   getRevisionPanel(): RevisionPanel {
     return this.revisions;
   }
@@ -577,10 +577,8 @@ export class InkswellView extends ItemView {
         break;
       case "plan": {
         const sub = this.subtab["plan"] ?? "overview";
-        if (sub === "board") this.board.render(panel);
-        else if (sub === "beats") this.beats.render(panel);
-        else if (sub === "grid") this.plotGrid.render(panel);
-        else if (sub === "outline") this.structure.render(panel);
+        if (sub === "beats") this.beats.render(panel);
+        else if (sub === "structure") this.structure.render(panel);
         else this.overview.render(panel);
         break;
       }
